Use createNextDescribe for define-class-fields
