+++ test/development/basic/define-class-fields/fixture/pages/animal.tsx
