+++ test/development/basic/define-class-fields/fixture/pages/derived.tsx
